+++ src/domain/protocol.rs
@@ -959,6 +959,26 @@ mod tests {
             protocol.text("My text Ґ Є №", Some(PageCode::PC1125)).unwrap(),
             &[77, 121, 32, 116, 101, 120, 116, 32, 0xF2, 32, 0xF4, 32, 0xFC]
         );
+        assert_eq!(
+            protocol.text("My text ‡ ť ¶", Some(PageCode::WPC1250)).unwrap(),
+            &[77, 121, 32, 116, 101, 120, 116, 32, 0x87, 32, 0x9D, 32, 0xB6]
+        );
+        assert_eq!(
+            protocol.text("My text Ђ џ Ъ", Some(PageCode::WPC1251)).unwrap(),
+            &[77, 121, 32, 116, 101, 120, 116, 32, 0x80, 32, 0x9F, 32, 0xDA]
+        );
+        assert_eq!(
+            protocol.text("My text ƒ ‰ ¥", Some(PageCode::WPC1253)).unwrap(),
+            &[77, 121, 32, 116, 101, 120, 116, 32, 0x83, 32, 0x89, 32, 0xA5]
+        );
+        assert_eq!(
+            protocol.text("My text Ğ Ş ş", Some(PageCode::WPC1254)).unwrap(),
+            &[77, 121, 32, 116, 101, 120, 116, 32, 0xD0, 32, 0xDE, 32, 0xFE]
+        );
+        assert_eq!(
+            protocol.text("My text Ђ Ә ғ", Some(PageCode::KZ1048)).unwrap(),
+            &[77, 121, 32, 116, 101, 120, 116, 32, 0x80, 32, 0xA3, 32, 0xBA]
+        );
 
         // With page code table not yet implemented
         assert!(protocol.text("My text", Some(PageCode::Hiragana)).is_err());